week4 readme update
