+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\HYUNA\.android\avd\360x640Q.avd" />
+            <value value="C:\Users\HYUNA\.android\avd\Pixel4R.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2021-11-11T20:00:04.109097100Z" />
+    <timeTargetWasSelectedWithDropDown value="2021-11-12T07:36:34.879049400Z" />
   </component>
 </project>
