+++ app/src/main/java/org/sopt/androidassignment1/ServiceCreator.kt
@@ -1,7 +1,5 @@
 package org.sopt.androidassignment1
 
-import org.sopt.androidassignment1.signIn.SigninService
-import org.sopt.androidassignment1.signup.SignupService
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
@@ -14,7 +12,5 @@ object ServiceCreator {
         .addConverterFactory(GsonConverterFactory.create())
         .build()
 
-    val signinService: SigninService = retrofit.create(SigninService::class.java)
-    val signupService: SignupService = retrofit.create(SignupService::class.java)
-
+    val SAMPLE_SERVICE: SampleService = retrofit.create(SampleService::class.java)
 }
