--- app/src/main/java/org/sopt/androidassignment1/signIn/SigninService.kt
@@ -1,16 +1,0 @@
-package org.sopt.androidassignment1.signIn
-
-import retrofit2.Call
-import retrofit2.http.Body
-import retrofit2.http.Headers
-import retrofit2.http.POST
-
-interface SigninService {
-    @Headers("Content-Type:application/json")
-    @POST("user/login")
-
-    fun postLogin(
-        @Body body : RequestSigninData
-    ) : Call<ResponseSigninData>
-
-}
